requirements.txt
pynput==1.7.7
pillow==10.4.0